requirements.txt
streamlit
plotly_express
pandas
numpy
